what?
Not sure why code didn't work. Error code:
Traceback (most recent call last):
  File "lightController.py", line 744, in <module>
    mode(strip)
  File "lightController.py", line 615, in bpm
    waitTime = 60 / int(BPM) # Calculate wait time based on BPM
ValueError: invalid literal for int() with base 10: '100.0'

+++ RGBserver/lightController.py
@@ -612,7 +612,7 @@ def bpm(strip):
                 return
             continue
 
-        waitTime = 60 / int(BPM) # Calculate wait time based on BPM
+        waitTime = 60 / float(BPM) # Calculate wait time based on BPM
         syncDelay = rawBPMdata["syncDelay"] # Check how long to wait for next beat in song
         requests.get(f"{serverAddress}/bpm?mode=resetDelay")
